pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncmy>=0.2.11",
    "cryptography>=47.0.0",
    "elasticsearch[async]>=8,<9",
    "fastapi[standard]>=0.136.1",
    "greenlet>=3.5.0",
    "jieba>=0.42.1",
    "langchain>=1.2.17",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "loguru>=0.7.3",
    "omegaconf>=2.3.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "sqlalchemy>=2.0.49",
]
